Revert parent component introduction
