+++ components/automate-ui/src/app/pages/integrations/azure-form/azure-form.component.ts
@@ -1,5 +1,6 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { BaseIntegrationFormComponent } from '../base-integration-form.component';
+import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
+import { FormGroup } from '@angular/forms';
+import { get } from 'lodash/fp';
 
 @Component({
   selector: 'app-azure-integration-form',
@@ -7,5 +8,12 @@ import { BaseIntegrationFormComponent } from '../base-integration-form.component
   styleUrls: ['./azure-form.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class IntegrationsAzureFormComponent extends BaseIntegrationFormComponent {
+export class IntegrationsAzureFormComponent {
+  @Input() formGroup: FormGroup;
+
+  showInstanceCreds() {
+    const formData = this.formGroup.value;
+    const no_creds = get('no_creds', formData);
+    return !no_creds;
+  }
 }
