--- components/automate-ui/src/app/pages/integrations/base-integration-form.component.ts
@@ -1,16 +1,0 @@
-import { Input, Component } from '@angular/core';
-import { FormGroup } from '@angular/forms';
-import { get } from 'lodash/fp';
-
-@Component({
-  template: ''
-})
-export class BaseIntegrationFormComponent {
-  @Input() formGroup: FormGroup;
-
-  showInstanceCreds() {
-    const formData = this.formGroup.value;
-    const no_creds = get('no_creds', formData);
-    return !no_creds;
-  }
-}
